Make dtoa output equivalent to that of nholmann/json
We use this for serializing f64 before sending in a JSON object to
spanner.

Internally spanner appears to use
zetasql (https://github.com/google/zetasql) which uses
nholmann/json (https://github.com/nlohmann/json/).

ZetaSQL will not accept a float in a json structure if it does not
roundtrip to the same value when serialized. By serializing the floats
using a grisu2 implementation equivalent to nholmann/json, we are
guaranteed that the float will be accepted.

You can test this out by pulling the zetasql repo and running:
```
bazel run //zetasql/tools/execute_query:execute_query \
    -- "select PARSE_JSON('2.250497700094831e+307');"
```